server/requirements.txt
fastapi
uvicorn
loguru
python-dotenv
ibm-watson
tweepy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib